# public/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "Fake Seam Connect", "version": "1.0.0" },
  "servers": [{ "url": "https://example.com" }],
...
  "paths": {
...
    "/access_codes/create": {
      "post": {
        "summary": "/access_codes/create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_code": {
                      "type": "object",
                      "properties": {
                        "access_code_id": { "type": "string" },
                        "device_id": { "type": "string" },
                        "name": { "type": "string" },
                        "code": { "type": "string" },
                        "created_at": { "type": "string" }
                      },
                      "required": [
                        "access_code_id",
                        "device_id",
                        "name",
                        "code",
                        "created_at"
                      ]
                    },
                    "ok": { "type": "boolean" }
                  },
                  "required": ["access_code", "ok"]
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "cst_ak_pk": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "device_id": { "type": "string" },
                  "name": { "type": "string" },
                  "code": { "type": "string" }
                },
                "required": ["device_id"]
              }
            }
          }
        },
        "tags": []
      },
      "get": {
        "summary": "/access_codes/create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_code": {
                      "type": "object",
                      "properties": {
                        "access_code_id": { "type": "string" },
                        "device_id": { "type": "string" },
                        "name": { "type": "string" },
                        "code": { "type": "string" },
                        "created_at": { "type": "string" }
                      },
                      "required": [
                        "access_code_id",
                        "device_id",
                        "name",
                        "code",
                        "created_at"
                      ]
                    },
                    "ok": { "type": "boolean" }
                  },
                  "required": ["access_code", "ok"]
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "cst_ak_pk": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "device_id": { "type": "string" },
                  "name": { "type": "string" },
                  "code": { "type": "string" }
                },
                "required": ["device_id"]
              }
            }
          }
        },
        "tags": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": { "type": "apiKey" },
      "cst_ak_pk": { "type": "apiKey" }
    }
  }
}